requirements.txt
pytest
requests
allure-pytest
assertpy
testfixtures